GAE/requirements.txt
Flask==2.3.2
line-bot-sdk==3.5.0
requests==2.31.0
google-cloud-secret-manager==2.16.2
gunicorn==20.1.0
pytz==2024.1